rTorrent: Allow linux screen launch
This pull requests add the `RT_DAEMON_MODE` environment variable which allows the user to configure how rTorrent is launched.

rTorrent is still launched in daemon mode by default unless the user changes it. When daemon mode is disabled, rTorrent will launch in the foreground on a linux screen instead.

This has some graceful exit benefits in limited circumstances.