rTorrent: Schedule torrent session save as background task
Addresses #281

This commit schedules the session save when a new download is inserted. The benefit if we can thread it as a background task. This avoids disruptions during startup and increases overall speed. The unique hash of the torrent is used as the schedule handle, to prevent accidently overwriting the task. It runs once and immediately.