rTorrent: Add disk space preallocation environment variable
rTorrent stickz introduces a new feature where it's possible to pre-allocate disk space for an entire torrent whether it's downloaded or not. The benefit to place files in sequence on the file system for faster seeding. Potential disk throughput on HDDs is increased due to reduced random access.

This pull request documents the different types of disk space pre-allocation for rTorrent and allows the user to configure it based on their needs. It's disabled by default because it does not benefit SSDs. The user should manually configure as required.