Upgrade to rTorrent v5.4
## Version 5.4 Release
This release includes two hash threads instead of one to better split the load between multiple threads. For instance, instead of 30% on a single thread it will execute two threads with 15%. Half the downloads are assigned to one thread and the other half to the other thread.

UDNS now works as intended. A build issue is fixed for debugging. A handshake build mount is added for later for multi-threading support.

## What's Changed
* libtorrent: Implement two hash queue threads by @stickz in https://github.com/stickz/rtorrent/pull/28
* libtorrent: Add handshake bind mount by @stickz in https://github.com/stickz/rtorrent/pull/29
* libtorrent: Add UDNS to configure.ac by @stickz in https://github.com/stickz/rtorrent/pull/32
* libtorrent: Fix instrumentation debug regression by @stickz in https://github.com/stickz/rtorrent/pull/33


**Full Changelog**: https://github.com/stickz/rtorrent/compare/v5.3-0.9.8-0.13.8...v5.4-0.9.8-0.13.8