rTorrent: Upgrade to v7.0
## Version 7.0 Release
This a major new release that provides the option to replace xmlrpc-c with tinyxml2. Tinyxml2 is 2 to 3 times faster for smaller responses and up to 30 times faster for large responses than it's xmlrpc-c counterpart. It is fully compatible with ruTorrent. It results in a significant reduction of in overhead, by eliminating the requirement for xmlrpc-c and it's stack of dependences.

To use simply configure rTorrent `--with-xmlrpc-tinyxml2` instead of `--with-xmlrpc-c`. This will be the default option in the near future!

## Tinyxml2 Benchmarks
xmlrpc-c
```
Benchmark               Time             CPU   Iterations
---------------------------------------------------------
small_response       3417 ns         3413 ns       204332
large_response   46159260 ns     46074301 ns           15
```

tinyxml2:
```
Benchmark               Time             CPU   Iterations
---------------------------------------------------------
small_response       1595 ns         1593 ns       462388
large_response    1512614 ns      1509299 ns          463
```

## What's Changed
* rTorrent: Add tinyxml2 support by stickz in https://github.com/stickz/rtorrent/pull/58
* tinyxml2: Fix commands without targets by stickz in https://github.com/stickz/rtorrent/pull/60

**Full Changelog**: https://github.com/stickz/rtorrent/compare/v6.3-0.9.8-0.13.8...v7.0-0.9.8-0.13.8