rTorrent: Switch to rakshasa v0.15.3
Support for rTorrent rakshasa has resumed. rTorrent stickz is being deprecated in favour of this new stable rakshasa release. It has tinyxml2 and UDNS bundled with the software. Requires #438 .

https://github.com/rakshasa/rtorrent/releases/tag/v0.15.3

The purpose of rtorrent stickz was due to dropped support of rtorrent rakshasa. It contained UDNS which was beneficial for stability. Now that rakshasa has resumed support, bundled UDNS and made a stable release, there is no reason left to maintain rtorrent stickz.